created the music list fragment and put the onlistitemclicked information in that class - MusicLibraryFragment and PlaylistFragment are now children of MusicListFragment and both the playlist and music library items are expandable
