+++ fanClub/src/com/lastcrusade/fanclub/components/MusicLibraryFragment.java
@@ -11,11 +11,10 @@ import android.content.ServiceConnection;
 import android.os.Bundle;
 import android.os.IBinder;
 import android.util.Log;
-import android.view.LayoutInflater;
 import android.view.View;
+import android.view.View.OnClickListener;
 import android.view.ViewGroup;
 
-import com.actionbarsherlock.app.SherlockListFragment;
 import com.lastcrusade.fanclub.CustomApp;
 import com.lastcrusade.fanclub.R;
 import com.lastcrusade.fanclub.model.SongMetadata;
@@ -23,10 +22,9 @@ import com.lastcrusade.fanclub.service.MusicLibraryService;
 import com.lastcrusade.fanclub.service.MusicLibraryService.MusicLibraryServiceBinder;
 import com.lastcrusade.fanclub.util.BroadcastRegistrar;
 import com.lastcrusade.fanclub.util.IBroadcastActionHandler;
-import com.lastcrusade.fanclub.util.ITitleable;
 import com.lastcrusade.fanclub.util.MusicListAdapter;
 
-public class MusicLibraryFragment extends SherlockListFragment implements ITitleable {
+public class MusicLibraryFragment extends MusicListFragment {
     private final String TAG = MusicLibraryFragment.class.getName();
     private BroadcastRegistrar registrar;
 
@@ -76,18 +74,7 @@ public class MusicLibraryFragment extends SherlockListFragment implements ITitle
         registerReceivers();
     }
     
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        super.onCreateView(inflater, container, savedInstanceState);
-        View v = inflater.inflate(R.layout.list, container, false);
-        return v;
-    }
-    
-    @Override
-    public void onResume(){
-        super.onResume();
-        getActivity().setTitle(getTitle());
-    }
-    
+        
     @Override
     public void onDestroy() {
         super.onDestroy();
@@ -124,6 +111,7 @@ public class MusicLibraryFragment extends SherlockListFragment implements ITitle
     private void unregisterReceivers() {
         this.registrar.unregister();
     }
+    
 
     private class MusicAdapter extends MusicListAdapter{
